Allow users to deactivate any 2FA recovery code directly from the UI.
1. **`ResponseDict`** : a TypedDict ensuring that all recovery code operations return a consistent dictionary structure.
2. **`_make_response`** : generates a standardized response dictionary for frontend alerts.
3. **`_process_recovery_code_response`** : processes recovery code operations and returns a structured response.
4. **`recovery_code_helper`** : a utility to perform generic recovery code operations (e.g., deactivate or delete) .

- `ResponseDict` enforces the structure of the dictionary returned by these functions, guaranteeing that frontend code receives consistent keys like `"SUCCESS"`, `"OPERATION_SUCCESS"`, `"TITLE"`, `"MESSAGE"`, and `"ALERT_TEXT"`.

These helper functions are then used by **`recovery_code_operation_helper`**, which takes in a request and a function to execute (e.g., `delete_code` or `invalidate_code`). Using the helper functions listed above, it returns a standardised JSON format that the frontend then uses to display messages through interactive popup alerts.

The `recovery_code_operation_helper` function also ensures that whenever a code is invalidated, it updates both the database and the cache. This guarantees that reads are pulled from the cache and the database is only accessed when changes occur.

**dashboard.js**

Modified the js function that controls the `invalidate code form` to use the standardised response returned from the backend to display whether a given action was successful(e.g deactivated) or whether the code has already been used, invalid or no longer exists.

**Workflow**

**Deactivated**
1. Login and navigate to the dashboard.
2. Generate raw codes.
3. Select any raw code and scroll down to the deactivate session.
4. Enter the raw code and press enter.
   → A success message will appear stating that the code has successfully been deactivated.

**Already deactivated**
5. Enter the same code again.
   → A blue info message will appear stating that the code has already been deactivated.

**Invalid code**
6. Enter an invalid code and press enter.
   → An error message will appear stating that the code is invalid.