**Feat:** Add `django_q` scheduler to automatically expire invalid codes
**Goal**

This commit introduces a background task scheduler using `django_q` to automatically delete expired or invalid codes. Currently, codes are only **soft deleted**, and this scheduler ensures they are removed in the background at a later time.

The scheduler can be configured in the admin dashboard to run at specified intervals: daily, weekly, monthly, quarterly, or yearly, with a set date and time. Once configured, it will automatically execute at the scheduled intervals.

**Implementation**

* Added a new model `RecoveryCodeCleanUpScheduler` to store scheduler configurations, including schedule type, run time, retention days, and logging preferences.
* Implemented a helper method `next_run_schedule()` to calculate the next execution time based on the schedule type.
* Overrode the model’s `save()` method to automatically populate `next_run` if it is not provided.
* Added a `ModelForm` with validation to ensure `next_run` can't be earlier than `run_at`, providing a user-friendly safeguard in the admin interface.
* Configured the Django admin to display `next_run` as a read-only field with dynamic help text explaining the next scheduled run based on the schedule type.

**Usage**

* Admins can create a scheduler instance and select the schedule type, date, and time.
* The scheduler will run automatically at the configured intervals and delete codes marked as invalid or expired.
* The system is designed to be safe and reusable, with built-in validations to prevent misconfiguration or database errors.

---

**apps.py**

Registered the scheduler in the app’s `ready()` method to ensure it is loaded **after all apps and models are initialised**. This prevents errors that would occur if the scheduler was to be  imported or started too early, before all the necessary models, signals, or functions are available.

---

**Models**

Added the following model:

* `RecoveryCodeAudit`

When the scheduler runs, it deletes the corresponding recovery codes and the batch model they belong to. While this action does clears up space, it also removes any historical records or auditing trail and this is where `RecoveryCodeAudit` steps in.

The `RecoveryCodeAudit` model addresses this by storing relevant information from the deleted models, providing an audit trail in a single model. The model records include:

* `action` – The action performed, e.g., invalidated or deleted
* `deleted_by` – Who deleted it, either a user or the system
* `user_issued_to` – The user the code was issued to
* `batch` – The batch the code belonged to
* `number_deleted` – The number of codes deleted
* `number_issued` – The number of codes issued
* `reason` – The reason for deletion, e.g., codes were expired
* `timestamp` – When the scheduler began the purge
* `updated_at` – Any updates made to the record

**Note:**
As an auditing model, `RecoveryCodeAudit` uses indexes on key fields to ensure queries are fast and efficient.

---

**signals.py**

Added two functions, `update_recovery_code_scheduler` and `validate_next_run_scheduler_value`, using `post_save` and `pre_save` signals:

* **`update_recovery_code_scheduler`**

  * Triggered via `post_save`.
  * When a scheduler instance is updated (via the admin UI or any view), this function automatically updates the scheduler.
  * This ensures changes take effect immediately without needing to restart the scheduler.

* **`validate_next_run_scheduler_value`**

  * Triggered via `pre_save`.
  * Runs before a scheduler instance is saved.
  * Ensures that the `next_run` value cannot be earlier than the `run_at` date if `next_run` is not `None`.

**Note:**

* If `next_run` is blank, it is automatically calculated based on the schedule type (daily, weekly, monthly, quarterly, or yearly).
* For example, if a daily schedule is set to run on 31/08/2025 at 16:30, `next_run` will be automatically set to 01/09/2025 at 16:30.

---

Added the function `purge_all_expired_batches`, which is executed by the `django_q` scheduler to purge expired and invalid codes.

The function also records the purge action using the `RecoveryCodePurgeHistory` model. This differs from `RecoveryCodeAudit` in that:

- `RecoveryCodePurgeHistory` only records the batches and the number of codes purged.
- `RecoveryCodeAudit` records detailed actions for each code within a batch, including numbers issued, who it was issued to, and other audit information.

Created a folder called `forms` and added a file `schedule_form.py`.

**schedule_form**

This form is used in the admin interface for the scheduler model. It provides a visual way to enforce that `run_at` cannot be later than `next_run` if `next_run` is not `None`.

**Note:**
- Although this validation is already enforced via signals, the form ensures that errors are displayed directly within the admin interface, improving usability.