This commit ensures that whenever a user generates a new batch of recovery codes via the UI, there is a cooldown period before a subsequent request can be made. If a user attempts to generate codes within the cooldown period, the wait time is incremented by using a multiplier to multiply the base cooldown, which becomes the new wait time. A cutoff point is also applied: if the wait time exceeds this threshold which then reset to the default cooldown period.
The behaviour can be customised via the following settings flags or via the admin:

- **`DJANGO_AUTH_RECOVERY_CODES_BASE_COOLDOWN`** – Base cooldown period (in seconds) before a user can generate another batch of recovery codes.
- **`DJANGO_AUTH_RECOVERY_CODES_COOLDOWN_MULTIPLIER`** – Multiplier applied to the base cooldown after each consecutive attempt within the cooldown.
- **`DJANGO_AUTH_RECOVERY_CODES_COOLDOWN_CUTOFF_POINT`** – Maximum cooldown period; if exceeded, the wait time resets to the default base cooldown.

- _can_codes_be_generated_yet(): Checks whether a code can be generated within the allowed time window. Returns True/False and the remaining wait time.
- _generate_code_batch(request, generate_with_expiry_date): Generates 2FA recovery codes based on the `generate_with_expiry_date` flag. If True, codes have an expiry date; if False, codes are indefinite.

These functions are used by `_generate_recovery_codes_with_expiry_date_helper` to generate recovery codes for the frontend. Depending on the request, it either:
- Generates codes with an expiry date,
- Generates codes without an expiry date, or
- Informs the user to wait before generating new codes.

This app makes extensive use of loggers, for example it uses `EmailSenderLogger` to record the sending of emails, and to record crucial information about the application's state. To facilitate this, a `loggers` folder was created containing two files: `loggers.py` and `logger_config.py`.

* **`loggers.py`**: Contains various loggers that can be imported by the app to log different states. Each logger has a distinct name, e.g. `email_sender_logger`, `purge_code_logger`, etc., allowing the app to log events to different files associated with each logger.

* **`logger_config.py`**: Contains the logger configuration, which performs two functions:

  1. Structures logging files by placing them inside a single folder.
  2. Provides a pre-configured `LOGGING` dictionary that can be imported directly into `settings.py`. Developers can simply import it without any setup:

```python
from django_auth_recovery_codes.loggers.logger_config import DJANGO_AUTH_RECOVERY_CODES_LOGGING

```

This approach allows developers to use the app with minimal configuration—plug and play.

Note:
The logger is optional if you don't import it the logger is not used
---

Integrated loggers in the `RecoveryCodesBatch` model to track crucial steps, particularly in areas where code generation might fail. This ensures that any issues during the generation of recovery codes are logged, making it easier to debug and monitor the application’s behaviour.

1. Log in and navigate to the dashboard.
2. Generate a code.
3. Attempt to generate a new code within the restricted time and a message will indicate that you must wait another X seconds.

**Note:** If the batch is deleted, a new batch can be generated, as the restriction only applies to an active batch.

Right now the batch displays in seconds will convert that to minutes and seconds to make it easier to read