The purpose of this commit is to display the time remaining before a new regeneration code can be requested in a  human-readable format (hours, minutes, or seconds). Previously, the remaining time was shown only in seconds.
- Added `converter.py` containing the `SecondsToTime` class.

`SecondsToTime` is a class that takes an integer value of seconds and converts it into a human-readable format.
For example, if the input is `3675` seconds, the output will be: `'1 hour, 1 minute and 15 seconds'`. This allows the UI to
display the remaining time in a format that is easy for the user to understand.