feat: Hide verification test form after successful setup
Ensure that the verification test form is removed from view once the user has successfully completed the verification process.

1. After the user generates their recovery codes and performs the verification test.
2. The form is first removed dynamically via JavaScript, and if the page is refreshed, the Jinja `{% if %}` statement ensures it remains hidden.

1. The fetch request sends the data to the `verify_test_code_setup` view function.
2. The view invokes the necessary classes to validate the test.
3. Once the test passes, the outcome e.g True or False (meaning the tests past or not) are stored in a timed-cache before it is used by the application to permanently hide the test verification form when the page is refreshed. The timed-cache stores the data for a specified duration before being removed, ensuring that each requests fetch data from the database only when necessary and only from the cache if no updates have been made. This keeps the data up to date while avoiding unnecessary database hits on page refresh.

The backend hides the verification form if the setup has not been completed, which requires a database check. Without caching, every page refresh would trigger a database query, which is inefficient as the data rarely changes. By storing the result in a timed cache after the first successful test, subsequent page refreshes read from the cache rather than the database. This allows the user to refresh the page multiple times without hitting the database. The duration the data is stored in the cache is determined by the relevant configuration flags.

= Added a classmethod called `has_first_time_set_up_occured` which takes a user and returns a Boolean indicating whether the user has run the tests for the first time. Once run the user cannot running it again even if they generate new batches.

Used cache to ensure that the page pulls from the cache after the set up has occurred.

Note:
This only stored in the cache if the tests were successful

1. Log in to the dashboard.
2. Generate recovery codes.
3. Enter a code in the verification field and press enter.
4. Observe that the outcome is displayed in the UI
5. The verification form will temporarily disappear when the tests pages;
6. Refresh the page and verification form will still not be visible.

Isssue

When the code are being tested the user can refresh the page which causes the tests to stop unlike when the code is being generated and the user is warned. Will add in the next commit.