Feat: Add a check to verify that the flags can only be "txt", "csv" or "pdf"
**Goal**

The goal of this commit is to provide a check that checks the format of how the code
can be downloaded as. The format the code can be downloaded as is either txt, csv or
pdf, (and it can be set in the settings) any other format will raise an error.

Added a function called `check_app_format_setting` which checks if th format is
valid. Raise a warning if which display in the terminal

Workflow (success)

1. In the `DJANGO_AUTH_RECOVERY_CODES_DEFAULT_FORMAT` flag set to either 'txt', 'csv', 'pdf'
2. A success message is display

Workflow (unsuccessful)

1. In the `DJANGO_AUTH_RECOVERY_CODES_DEFAULT_FORMAT` flag set to either 'text'

A warning message is displayed

```
WARNINGS:
?: DJANGO_AUTH_RECOVERY_CODES_DEFAULT_FORMAT
        HINT: The format must be either pdf, txt or csv but got text
```