Feat: Allow users to log out from the 2FA dashboard
*Goal*
This commit introduces the ability for users to log out via the 2FA dashboard interface.  After logout, users are redirected according to the `DJANGO_AUTH_RECOVERY_CODE_REDIRECT_VIEW` setting.

- If a valid view name is provided, the user is redirected to that view.
- If the view name is invalid or not provided, the user is redirected to the "/" page.

Note:
As this is a reusable application, it does not assume the presence of a specific home page.  The default redirect to "/" assumes that the host application has created a page at this route.

**conf.py**

Added a flag check using the file dictionary.
This allows developers using the application to run `python manage.py check` to verify whether the value provided for the `DJANGO_AUTH_RECOVERY_CODE_REDIRECT_VIEW` setting is valid.

**urls.py**

Added a URL to enable the logout button to log a user out when clicked.

**views.py**

Created a logout view that points the logout url to the view which when click logs the user out.

**sidebar.html**

Added a logout form containing a CSRF token to enhance security when users log out.

**Workflow:**

1. Log in and navigate to the dashboard.
2. Add the flag "DJANGO_AUTH_RECOVERY_CODE_REDIRECT_VIEW" in your project and point to a view e.g "/" (home page)
3. Now click log out
4. You will be redirect the home page "/"